claude/examples/basic_project/requirements.txt
# Basic Django Project Requirements
# Core dependencies for a multi-tenant Django application

# Django framework
Django==5.2.4
django-environ==0.12.0

# Database
psycopg2-binary==2.9.10

# Image processing
pillow==10.4.0

# Forms
django-crispy-forms==2.3
crispy-bootstrap5==2024.10

# Production server
gunicorn==23.0.0

# REST API (optional - remove if not building APIs)
djangorestframework==3.15.2
